ref(data): implementing versions for real
consolidated components and data sub-packages into data to avoid circular deps

data.AvailableVersionsFromAPI{} is an in-progress implementation of the current versions API multi-get endpoint, tests needed
